VST sync patch: compatibility fix for 64bit builds
It seems 64bit builds for some reason have problems with VST Sync feature on,
workaround seems to be converting VST sync patch from double to floats,
which does work both with 32 and 64bit builds. Double precision
seems to produce odd numbers with 64bit build. (tested on VirtualBox Linux
Mint 14.1 64 bit OS)
(cherry picked from commit 011f87e6e60cccd16f3783e9c4885e03d95c1e56)

+++ include/VST_sync_shm.h
@@ -39,19 +39,19 @@
 struct sncVST
 {
 	bool isPlayin;
-	double ppqPos;
+	float ppqPos;
 	int timeSigNumer;
 	int timeSigDenom;
 	bool isCycle;
 	bool hasSHM;
-	double cycleStart;
-	double cycleEnd;
+	float cycleStart;
+	float cycleEnd;
 	int m_bufferSize;
 	int m_sampleRate;
 	int m_bpm;
 
 #ifdef VST_SNC_LATENCY
-	double m_latency;
+	float m_latency;
 #endif
 } ;
